+++ NEWS.md
@@ -1,3 +1,9 @@
+### purgeR 1.2
+
+- Fix computation of total opportunity of purging (*O*).
+- Function `ip_op()` only computes the expressed opportunity of purging (*Oe*) by default.
+- Add new function `ped_sort()` to sort pedigrees.
+
 ### purgeR 1.1
 
 - Fixed issue in `pop_Nancestors` due to bad access to index.